+++ src/main/java/uk/ac/brighton/uni/ch629/ecengine/component/ClickableComponent.java
@@ -1,17 +1,15 @@
 package uk.ac.brighton.uni.ch629.ecengine.component;
 
 import uk.ac.brighton.uni.ch629.ecengine.entity.Entity;
-import uk.ac.brighton.uni.ch629.ecengine.logic.World;
 
 import java.awt.*;
-import java.util.UUID;
 
 public class ClickableComponent extends Component {
     public ClickableComponent(Entity parent) {
         super(parent);
     }
 
-    public void update(int deltaTime) {
+    public void update(double deltaTime) {
 
     }
